Init Auth

+++ src/main/kotlin/com/example/projet_tutore_back/data/repository/MessageRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface MessageRepository : CrudRepository<Message, Int> {
+interface MessageRepository : CrudRepository<Message, Long> {
     fun save(message: Message) : Message
 }
